Move PluggableCommandLineOption to plugincode #2233
This breaks circular dependencies between plugincode and scancode
