+++ src/cluecode/plugin_filter_clues.py
@@ -32,10 +32,10 @@ from itertools import chain
 import attr
 from six import string_types
 
+from plugincode import PluggableCommandLineOption
+from plugincode import POST_SCAN_GROUP
 from plugincode.post_scan import PostScanPlugin
 from plugincode.post_scan import post_scan_impl
-from scancode import PluggableCommandLineOption
-from scancode import POST_SCAN_GROUP
 
 
 def logger_debug(*args):
