+++ src/formattedcode/output_json.py
@@ -33,10 +33,10 @@ from six import string_types
 from commoncode.system import py2
 from commoncode.system import py3
 from formattedcode import FileOptionType
+from plugincode import PluggableCommandLineOption
+from plugincode import OUTPUT_GROUP
 from plugincode.output import output_impl
 from plugincode.output import OutputPlugin
-from scancode import PluggableCommandLineOption
-from scancode import OUTPUT_GROUP
 
 
 """
