+++ src/formattedcode/output_jsonlines.py
@@ -32,10 +32,10 @@ import simplejson
 from commoncode.system import py2
 from commoncode.system import py3
 from formattedcode import FileOptionType
-from plugincode.output import output_impl
+from plugincode import OUTPUT_GROUP
+from plugincode import PluggableCommandLineOption
 from plugincode.output import OutputPlugin
-from scancode import PluggableCommandLineOption
-from scancode import OUTPUT_GROUP
+from plugincode.output import output_impl
 
 """
 Output plugin to write scan results as JSON lines.
