+++ src/scancode/plugin_ignore.py
@@ -30,8 +30,8 @@ from functools import partial
 from commoncode.fileset import is_included
 from plugincode.pre_scan import PreScanPlugin
 from plugincode.pre_scan import pre_scan_impl
-from scancode import PluggableCommandLineOption
-from scancode import PRE_SCAN_GROUP
+from plugincode import PluggableCommandLineOption
+from plugincode import PRE_SCAN_GROUP
 from commoncode import compat
