+++ src/summarycode/plugin_consolidate.py
@@ -28,7 +28,6 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from collections import Counter
-from collections import defaultdict
 from collections import OrderedDict
 
 import attr
@@ -42,8 +41,8 @@ from packagedcode.build import BaseBuildManifestPackage
 from packagedcode.utils import combine_expressions
 from plugincode.post_scan import PostScanPlugin
 from plugincode.post_scan import post_scan_impl
-from scancode import PluggableCommandLineOption
-from scancode import POST_SCAN_GROUP
+from plugincode import PluggableCommandLineOption
+from plugincode import POST_SCAN_GROUP
 from summarycode import copyright_summary
